Add support for an additional elmo controller for the torso
This adds a parameter: "torso_present". If active, the node searches for a 5th elmo
controller on the bus, and configures it to velocity mode.

It will listen to a joint in a joint_state message (giskardCommand).
The joint is called "triangle_base_joint".

It implements its own 1DOF interpolator for the torso, plus an independent watchdog.